+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - [Autocompletion](./README.md#autocompletion) for URL protocols in the MediaWiki mode
 - [CodeFolding](./README.md#codefolding) for extension tags with specified modes in the MediaWiki mode
-- Highlighting of redirection in the MediaWiki mode
+- Highlighting of redirection and image parameters in the MediaWiki mode
 
 **Fixed**